+++ NEWS.md
@@ -1,3 +1,8 @@
+# truelies 0.2.0
+
+* Added `empirical_bayes()` function.
+* `update_prior()` now uses `stats::dunif` as the prior by default.
+
 # truelies 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.